+++ minerva_analysis/client/.gitignore
@@ -1,4 +1,5 @@
 data/
+coverage/
 .editorconfig
 .eslintignore
 .eslintrc.json
