+++ minerva_analysis/client/package.json
@@ -5,7 +5,9 @@
   "scripts": {
     "start": "webpack",
     "watch": "webpack --watch",
-    "test": "karma start karma.conf.js"
+    "test:test": "tsc --project test/tsconfig.json",
+    "test:src": "mockttp -c karma start karma.conf.js",
+    "test": "npm run test:test && npm run test:src"
   },
   "license": "MIT",
   "dependencies": {
@@ -22,6 +24,7 @@
     "install": "^0.13.0",
     "jquery": "^3.5.1",
     "jquery-form": "^4.3.0",
+    "karma-sourcemap-loader": "^0.3.8",
     "lensing": "^1.0.17",
     "lodash": "^4.17.19",
     "mark.js": "^8.11.1",
@@ -34,11 +37,16 @@
     "viawebgl": "github:thejohnhoffer/viaWebGL#webgl2"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.6",
+    "@babel/core": "^7.17.12",
     "@babel/plugin-transform-runtime": "^7.12.1",
-    "@babel/preset-env": "^7.11.5",
+    "@babel/preset-env": "^7.17.12",
+    "@babel/preset-typescript": "^7.17.12",
     "@babel/runtime": "^7.17.9",
     "@babel/runtime-corejs2": "^7.17.11",
+    "@httptoolkit/proxy-agent": "^5.0.1-socks-lookup-fix.0",
+    "@types/chai": "^4.3.1",
+    "@types/karma-fixture": "^0.2.5",
+    "@types/mocha": "^9.1.1",
     "babel-loader": "^8.1.0",
     "chai": "^4.3.6",
     "karma": "^6.3.19",
@@ -51,6 +59,8 @@
     "karma-mocha": "^2.0.1",
     "karma-webpack": "^5.0.0",
     "mocha": "^10.0.0",
+    "mockttp": "^3.0.0",
+    "typescript": "^4.6.4",
     "webpack": "^4.44.2",
     "webpack-cli": "^3.3.12"
   }
